feat(vtkImageMapper): support coincident topology
