+++ Sources/Rendering/OpenGL/PolyDataMapper/index.js
@@ -11,6 +11,8 @@ import vtkViewNode from 'vtk.js/Sources/Rendering/SceneGraph/ViewNode';
 import vtkPolyDataVS from 'vtk.js/Sources/Rendering/OpenGL/glsl/vtkPolyDataVS.glsl';
 import vtkPolyDataFS from 'vtk.js/Sources/Rendering/OpenGL/glsl/vtkPolyDataFS.glsl';
 
+import vtkReplacementShaderMapper from 'vtk.js/Sources/Rendering/OpenGL/ReplacementShaderMapper';
+
 /* eslint-disable no-lonely-if */
 
 const primTypes = {
@@ -919,75 +921,6 @@ function vtkOpenGLPolyDataMapper(publicAPI, model) {
     //   return;
     // }
     return cp;
-  };
-
-  publicAPI.replaceShaderCoincidentOffset = (shaders, ren, actor) => {
-    const cp = publicAPI.getCoincidentParameters(ren, actor);
-
-    // if we need an offset handle it here
-    // The value of .000016 is suitable for depth buffers
-    // of at least 16 bit depth. We do not query the depth
-    // right now because we would need some mechanism to
-    // cache the result taking into account FBO changes etc.
-    if (cp && (cp.factor !== 0.0 || cp.offset !== 0.0)) {
-      let FSSource = shaders.Fragment;
-
-      FSSource = vtkShaderProgram.substitute(
-        FSSource,
-        '//VTK::Coincident::Dec',
-        ['uniform float cfactor;', 'uniform float coffset;']
-      ).result;
-
-      if (model.context.getExtension('EXT_frag_depth')) {
-        if (cp.factor !== 0.0) {
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::UniformFlow::Impl',
-            [
-              'float cscale = length(vec2(dFdx(gl_FragCoord.z),dFdy(gl_FragCoord.z)));',
-              '//VTK::UniformFlow::Impl',
-            ],
-            false
-          ).result;
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::Depth::Impl',
-            'gl_FragDepthEXT = gl_FragCoord.z + cfactor*cscale + 0.000016*coffset;'
-          ).result;
-        } else {
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::Depth::Impl',
-            'gl_FragDepthEXT = gl_FragCoord.z + 0.000016*coffset;'
-          ).result;
-        }
-      }
-      if (model.openGLRenderWindow.getWebgl2()) {
-        if (cp.factor !== 0.0) {
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::UniformFlow::Impl',
-            [
-              'float cscale = length(vec2(dFdx(gl_FragCoord.z),dFdy(gl_FragCoord.z)));',
-              '//VTK::UniformFlow::Impl',
-            ],
-            false
-          ).result;
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::Depth::Impl',
-            'gl_FragDepth = gl_FragCoord.z + cfactor*cscale + 0.000016*coffset;'
-          ).result;
-        } else {
-          FSSource = vtkShaderProgram.substitute(
-            FSSource,
-            '//VTK::Depth::Impl',
-            'gl_FragDepth = gl_FragCoord.z + 0.000016*coffset;'
-          ).result;
-        }
-      }
-      shaders.Fragment = FSSource;
-    }
   };
 
   publicAPI.replaceShaderPicking = (shaders, ren, actor) => {
@@ -1969,6 +1902,11 @@ export function extend(publicAPI, model, initialValues = {}) {
 
   // Inheritance
   vtkViewNode.extend(publicAPI, model, initialValues);
+  vtkReplacementShaderMapper.implementReplaceShaderCoincidentOffset(
+    publicAPI,
+    model,
+    initialValues
+  );
 
   model.primitives = [];
   model.primTypes = primTypes;
